Multiple days work, check description
Changed add command, now you have one and change the asked type instead of one for each object type.
Added GUI for Decision object, Direction object.
Made key groups, so you can filter which keys are allowed.
Added the option of a variable to the Decision object, so the statement is a variable string, you can execute with eval.
You can check if a variable exists in the robot with a function, if it does you get the value of the variable.